+++ app/src/main/res/layout/fragment_main_event1.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.MainEvent1">
+    tools:context=".Event.MainEvent1">
 
     <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/iv_event1"
