+++ app/src/main/res/layout/fragment_main_event3.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".View.MainEvent3">
+    tools:context=".Event.MainEvent3">
 
     <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/iv_event3"
